apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "xiaohei-erp-api"
version = "0.1.0"
description = "Minimal FastAPI backend for Xiaohei ERP"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "httpx>=0.28,<1.0",
  "psycopg[binary]>=3.2,<4.0",
  "pydantic-settings>=2.7,<3.0",
  "SQLAlchemy>=2.0,<3.0",
  "uvicorn>=0.34,<1.0"
]

[project.optional-dependencies]
dev = [
]
